Refactor large files
Refactor all large files in the project.
