+++ src/components/AddEvacuationPointDialog.tsx
@@ -1,3 +1,4 @@
+
 import React, { useState } from 'react';
 import { useForm } from 'react-hook-form';
 import { zodResolver } from '@hookform/resolvers/zod';
@@ -11,45 +12,21 @@ import {
   DialogDescription,
   DialogFooter,
 } from "@/components/ui/dialog";
-import {
-  Form,
-  FormControl,
-  FormField,
-  FormItem,
-  FormLabel,
-  FormMessage,
-} from "@/components/ui/form";
-import { Input } from "@/components/ui/input";
-import { Textarea } from "@/components/ui/textarea";
-import { Checkbox } from "@/components/ui/checkbox";
 import { EvacuationPoint } from '../types/emergency';
-import { ScrollArea } from '@/components/ui/scroll-area';
 import { toast } from '@/components/ui/use-toast';
-import { Locate } from "lucide-react";
+import { EvacuationPointForm, evacuationPointFormSchema } from './EvacuationPointForm';
+import { ActionChoiceModal } from './ActionChoiceModal';
+import { submitProposal, EvacPointWithContact } from '../services/evacuationPointService';
 
-// ---- Tipado para el borrador local (id, creadoPor extendido, etc)
 type LocalEvacPoint = Omit<EvacuationPoint, 'id' | 'status' | 'restrictions'> & {
   id: string;
   status?: string;
   restrictions?: string;
+  email?: string;
+  phone?: string;
 };
 
-// --- NUEVO: tipo auxiliar para incluir datos de contacto
-type EvacPointWithContact = Omit<EvacuationPoint, 'id'> & { email?: string; phone?: string };
-
 const LOCAL_STORAGE_KEY = "evacuation-point-borradores-v2";
-
-const formSchema = z.object({
-  name: z.string().min(1, "El nombre es obligatorio."),
-  locality: z.string().min(1, "La localidad es obligatoria."),
-  description: z.string().optional(),
-  lat: z.coerce.number().min(-90, "Latitud inválida").max(90, "Latitud inválida"),
-  lng: z.coerce.number().min(-180, "Longitud inválida").max(180, "Longitud inválida"),
-  createdBy: z.string().min(1, "El autor es obligatorio."),
-  isDaytimeOnly: z.boolean().default(false),
-  email: z.string().email("Email inválido").or(z.literal("")).optional(),
-  phone: z.string().min(7, "Teléfono inválido").or(z.literal("")).optional(),
-});
 
 interface AddEvacuationPointDialogProps {
   open: boolean;
@@ -63,9 +40,10 @@ export const AddEvacuationPointDialog: React.FC<AddEvacuationPointDialogProps> =
   const [actionModal, setActionModal] = useState(false);
   const [pendingData, setPendingData] = useState<EvacPointWithContact | null>(null);
   const [loading, setLoading] = useState(false);
+  const [formError, setFormError] = useState("");
 
-  const form = useForm<z.infer<typeof formSchema>>({
-    resolver: zodResolver(formSchema),
+  const form = useForm<z.infer<typeof evacuationPointFormSchema>>({
+    resolver: zodResolver(evacuationPointFormSchema),
     defaultValues: {
       name: '',
       locality: '',
@@ -74,20 +52,20 @@ export const AddEvacuationPointDialog: React.FC<AddEvacuationPointDialogProps> =
       lng: 0,
       createdBy: '',
       isDaytimeOnly: false,
+      email: '',
+      phone: '',
     },
   });
 
-  // --- Cargar borradores de localStorage al abrir modal sólo la primera vez
   React.useEffect(() => {
     if (open) {
       try {
         const raw = localStorage.getItem(LOCAL_STORAGE_KEY);
         if (raw) setBorradores(JSON.parse(raw));
-      } catch { /* vacío */ }
+      } catch { /* no-op */ }
     }
   }, [open]);
 
-  // --- Guardar borradores en localStorage
   function saveBorradores(newBorradores: LocalEvacPoint[]) {
     localStorage.setItem(LOCAL_STORAGE_KEY, JSON.stringify(newBorradores));
     setBorradores(newBorradores);
@@ -107,46 +85,17 @@ export const AddEvacuationPointDialog: React.FC<AddEvacuationPointDialogProps> =
       description: "Puedes consultarlo y editarlo desde este navegador."
     });
   }
+
   function eliminarBorrador(id: string) {
     const nuevos = borradores.filter(b => b.id !== id);
     saveBorradores(nuevos);
     toast({ title: "Borrador eliminado" });
   }
 
-  // --- Envío remoto
-  async function enviarPropuesta(
-    datos: EvacPointWithContact,
-    quitarLocalId?: string
-  ) {
+  async function enviarPropuesta(datos: EvacPointWithContact, quitarLocalId?: string) {
     setLoading(true);
     try {
-      const currentUser = { email: datos.email || "usuario@ejemplo.com" };
-      await fetch(
-        "https://script.google.com/macros/s/AKfycbx1gOPeFd59uKv6FQSG5hFYxzKNcHFs-Xxoiwvmd4ZdQBXogmpkxox3nq4zIVjSnoCBpA/exec",
-        {
-          method: "POST",
-          headers: { "Content-Type": "application/json" },
-          body: JSON.stringify({
-            usuario: currentUser.email,
-            telefono: datos.phone || "",
-            datos: {
-              nombre: datos.createdBy,
-              telefono: datos.phone || "",
-              puntoDeAterrizaje: {
-                nombre: datos.name,
-                localidad: datos.locality,
-                latitud: datos.lat,
-                longitud: datos.lng,
-                descripcion: datos.description,
-                restricciones: datos.restrictions || '',
-                soloDiurno: datos.isDaytimeOnly,
-                observaciones: "",
-                fotos: datos.photos ?? [],
-              },
-            },
-          }),
-        }
-      );
+      await submitProposal(datos);
       toast({
         title: "Propuesta enviada",
         description: "Tu propuesta ha sido enviada para revisión.",
@@ -184,40 +133,8 @@ export const AddEvacuationPointDialog: React.FC<AddEvacuationPointDialogProps> =
     }
   };
 
-  // --- NUEVA función para ubicar al usuario
-  const handleLocateMe = () => {
-    if (!navigator.geolocation) {
-      toast({
-        variant: "destructive",
-        title: "Geolocalización no soportada",
-        description: "Tu navegador no admite la geolocalización."
-      });
-      return;
-    }
-    toast({ title: "Buscando tu ubicación..." });
-    navigator.geolocation.getCurrentPosition(
-      (pos) => {
-        const { latitude, longitude } = pos.coords;
-        form.setValue("lat", latitude);
-        form.setValue("lng", longitude);
-        toast({
-          title: "Ubicación detectada",
-          description: `Latitud: ${latitude.toFixed(5)}, Longitud: ${longitude.toFixed(5)}`
-        });
-      },
-      (err) => {
-        toast({
-          variant: "destructive",
-          title: "No se pudo obtener la ubicación",
-          description: err.message || "Verifica permisos y conexión."
-        });
-      }
-    );
-  };
-
-  // --- Cuando el usuario pulse "Guardar Punto", mostrar elección modal
-  const onSubmit = (values: z.infer<typeof formSchema>) => {
-    setFormError(""); // Limpiar error previo
+  const onSubmit = (values: z.infer<typeof evacuationPointFormSchema>) => {
+    setFormError("");
     const pointData: EvacPointWithContact = {
       name: values.name,
       locality: values.locality,
@@ -232,12 +149,10 @@ export const AddEvacuationPointDialog: React.FC<AddEvacuationPointDialogProps> =
       email: values.email,
       phone: values.phone,
     };
-    setPendingData(pointData); // Almacena datos para la siguiente acción
-    setActionModal(true); // Abre modal de elección
+    setPendingData(pointData);
+    setActionModal(true);
   };
-  const [formError, setFormError] = useState("");
-
-  // --- Después de completar acción, limpiar todo
+  
   const resetAll = () => {
     form.reset();
     setPhotos([]);
@@ -246,77 +161,25 @@ export const AddEvacuationPointDialog: React.FC<AddEvacuationPointDialogProps> =
     onOpenChange(false);
   };
 
-  // --- Modal de elección de acción: guardar, enviar, ambas
-  function ActionChoiceModal() {
-    if (!actionModal || !pendingData) return null;
-    // Lógica para controlar error al compartir
-    const compartir = async (guardarTambien: boolean) => {
-      setFormError("");
-      if (
-        (!pendingData.email || pendingData.email.trim() === "") &&
-        (!pendingData.phone || pendingData.phone.trim() === "")
-      ) {
-        setFormError("Para compartir para revisión, por favor indica al menos un correo electrónico o teléfono de contacto.");
-        return;
-      }
-      if (guardarTambien) {
-        guardarBorrador(pendingData);
-      }
-      await enviarPropuesta(pendingData);
-      resetAll();
-    };
-    return (
-      <div className="fixed inset-0 bg-black/40 flex justify-center items-center z-30">
-        <div className="bg-white p-4 rounded-xl shadow-lg max-w-xs w-full text-center">
-          <h3 className="font-bold mb-2">¿Qué deseas hacer?</h3>
-          <p className="text-sm mb-3">
-            Puedes guardar el formulario en tu dispositivo, compartirlo para revisión o ambas cosas.<br />
-            <span className="text-xs text-muted-foreground block mt-2">
-              <strong>Aviso de privacidad:</strong> Tu email o teléfono solo se usarán para verificar y contactar en caso necesario sobre la propuesta, nunca para uso comercial.
-            </span>
-          </p>
-          {formError && (
-            <div className="mb-2 text-destructive text-sm">{formError}</div>
-          )}
-          <div className="flex flex-col gap-2">
-            <Button
-              variant="default"
-              disabled={loading}
-              onClick={() => { guardarBorrador(pendingData); resetAll(); }}
-            >
-              Guardar sólo en mi dispositivo
-            </Button>
-            <Button
-              disabled={loading}
-              onClick={() => compartir(false)}
-            >
-              Compartir sólo para revisión
-            </Button>
-            <Button
-              disabled={loading}
-              onClick={() => compartir(true)}
-            >
-              Guardar y compartir ambas cosas
-            </Button>
-            <Button variant="outline" onClick={resetAll}>
-              Cancelar
-            </Button>
-          </div>
-        </div>
-      </div>
-    );
-  }
-
-  const handleClose = () => {
-    form.reset();
-    setPhotos([]);
-    setPendingData(null);
-    setActionModal(false);
-    onOpenChange(false);
+  const compartir = async (guardarTambien: boolean) => {
+    if (!pendingData) return;
+    setFormError("");
+    if (
+      (!pendingData.email || pendingData.email.trim() === "") &&
+      (!pendingData.phone || pendingData.phone.trim() === "")
+    ) {
+      setFormError("Para compartir para revisión, por favor indica al menos un correo electrónico o teléfono de contacto.");
+      return;
+    }
+    if (guardarTambien) {
+      guardarBorrador(pendingData);
+    }
+    await enviarPropuesta(pendingData);
+    resetAll();
   };
 
   return (
-    <Dialog open={open} onOpenChange={handleClose}>
+    <Dialog open={open} onOpenChange={resetAll}>
       <DialogContent className="sm:max-w-[600px]">
         <DialogHeader>
           <DialogTitle>Añadir/Proponer Punto de Aterrizaje</DialogTitle>
@@ -324,122 +187,27 @@ export const AddEvacuationPointDialog: React.FC<AddEvacuationPointDialogProps> =
             Introduce los detalles del nuevo punto.
           </DialogDescription>
         </DialogHeader>
-        <Form {...form}>
-          <form onSubmit={form.handleSubmit(onSubmit)}>
-            <ScrollArea className="h-[60vh] pr-6">
-              <div className="space-y-4 py-4">
-                <FormField control={form.control} name="name" render={({ field }) => (
-                  <FormItem>
-                    <FormLabel>Nombre</FormLabel>
-                    <FormControl><Input placeholder="Ej: Campo de fútbol de Arnedo" {...field} /></FormControl>
-                    <FormMessage />
-                  </FormItem>
-                )}/>
-                <FormField control={form.control} name="locality" render={({ field }) => (
-                  <FormItem>
-                    <FormLabel>Localidad</FormLabel>
-                    <FormControl><Input placeholder="Ej: Arnedo" {...field} /></FormControl>
-                    <FormMessage />
-                  </FormItem>
-                )}/>
-                <div className="grid grid-cols-2 gap-4 items-end">
-                  <FormField control={form.control} name="lat" render={({ field }) => (
-                    <FormItem>
-                      <FormLabel>Latitud</FormLabel>
-                      <FormControl>
-                        <Input type="number" placeholder="42.2245" {...field} />
-                      </FormControl>
-                      <FormMessage />
-                    </FormItem>
-                  )}/>
-                  <div className="flex flex-row gap-2 items-end">
-                    <FormField control={form.control} name="lng" render={({ field }) => (
-                      <FormItem className="flex-1">
-                        <FormLabel>Longitud</FormLabel>
-                        <FormControl>
-                          <Input type="number" placeholder="-2.1018" {...field} />
-                        </FormControl>
-                        <FormMessage />
-                      </FormItem>
-                    )}/>
-                    {/* Botón para obtener ubicación */}
-                    <button
-                      type="button"
-                      aria-label="Obtener mi ubicación"
-                      className="h-10 w-10 bg-secondary flex items-center justify-center rounded-md border ml-1 hover:bg-accent transition-colors"
-                      onClick={handleLocateMe}
-                      title="Ubicarme automáticamente"
-                    >
-                      <Locate className="text-primary" />
-                    </button>
-                  </div>
-                </div>
-                <FormField control={form.control} name="description" render={({ field }) => (
-                  <FormItem>
-                    <FormLabel>Descripción</FormLabel>
-                    <FormControl><Textarea placeholder="Terreno llano, con acceso por carretera..." {...field} /></FormControl>
-                    <FormMessage />
-                  </FormItem>
-                )}/>
-                <FormField control={form.control} name="createdBy" render={({ field }) => (
-                  <FormItem>
-                    <FormLabel>Creado por</FormLabel>
-                    <FormControl><Input placeholder="Nombre del agente" {...field} /></FormControl>
-                    <FormMessage />
-                  </FormItem>
-                )}/>
-                {/* CONTACTO */}
-                <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
-                  <FormField control={form.control} name="email" render={({ field }) => (
-                    <FormItem>
-                      <FormLabel>Email de contacto</FormLabel>
-                      <FormControl><Input type="email" placeholder="correo@ejemplo.com" {...field} /></FormControl>
-                      <FormMessage />
-                    </FormItem>
-                  )}/>
-                  <FormField control={form.control} name="phone" render={({ field }) => (
-                    <FormItem>
-                      <FormLabel>Teléfono de contacto</FormLabel>
-                      <FormControl><Input type="tel" placeholder="Ej: 612345678" {...field} /></FormControl>
-                      <FormMessage />
-                    </FormItem>
-                  )}/>
-                </div>
-                <span className="block text-xs text-muted-foreground mt-[-10px] mb-2">
-                  {/* Aclaración sobre protección de datos */}
-                  El email o teléfono solo se usarán para comprobar la veracidad de la propuesta.
-                </span>
-                <FormItem>
-                  <FormLabel>Fotos de referencia</FormLabel>
-                  <FormControl>
-                    <Input type="file" multiple accept="image/*" onChange={handlePhotoUpload} />
-                  </FormControl>
-                  <div className="grid grid-cols-3 gap-2 mt-2">
-                    {photos.map((photo, index) => (
-                      <img key={index} src={photo} alt={`preview ${index}`} className="rounded-md object-cover h-24 w-full" />
-                    ))}
-                  </div>
-                </FormItem>
-                <FormField control={form.control} name="isDaytimeOnly" render={({ field }) => (
-                  <FormItem className="flex flex-row items-start space-x-3 space-y-0 rounded-md border p-4">
-                    <FormControl>
-                      <Checkbox checked={field.value} onCheckedChange={field.onChange} />
-                    </FormControl>
-                    <div className="space-y-1 leading-none">
-                      <FormLabel>Solo para uso diurno</FormLabel>
-                    </div>
-                  </FormItem>
-                )}/>
-              </div>
-            </ScrollArea>
-            <DialogFooter className="pt-4">
-              <Button variant="outline" type="button" onClick={handleClose}>Cancelar</Button>
-              <Button type="submit">Guardar Punto</Button>
-            </DialogFooter>
-          </form>
-          {/* Opciones de acción al guardar */}
-          <ActionChoiceModal />
-        </Form>
+        <EvacuationPointForm 
+          form={form} 
+          onSubmit={onSubmit}
+          photos={photos}
+          onPhotoUpload={handlePhotoUpload}
+        >
+          <DialogFooter className="pt-4 border-t">
+            <Button variant="outline" type="button" onClick={resetAll}>Cancelar</Button>
+            <Button type="submit">Guardar Punto</Button>
+          </DialogFooter>
+        </EvacuationPointForm>
+        
+        <ActionChoiceModal
+          open={actionModal}
+          loading={loading}
+          formError={formError}
+          onSaveOnly={() => { if (pendingData) { guardarBorrador(pendingData); resetAll(); } }}
+          onShareOnly={() => compartir(false)}
+          onSaveAndShare={() => compartir(true)}
+          onCancel={resetAll}
+        />
       </DialogContent>
     </Dialog>
   );
